Revamp (backend)
I have more or less just finished my programming course (backend with webservers) and when
I looked at this code, it was shit.
Thus I decided to change a lot of stuff (not a lot, but making the code more readable and
comprehensible). So I have reconsidered how the functions should function and to be utilized,
better logic (could probably be improved) and better comments.